refactor: 홈 화면 유저 정보 조회를 profile 컴포넌트로 이동
